toplist history and suggestion delete just the suggstion for one user

+++ install/update-10.sql
@@ -1,4 +1,4 @@
-ALTER TABLE user ADD COLUMN suggestion_up_to_date INTEGER;
+ALTER TABLE user ADD COLUMN suggestion_up_to_date INTEGER DEFAULT 0;
 
 update user set suggestion_up_to_date = 0;
 
@@ -46,11 +46,11 @@ BEGIN
          START TRANSACTION;
             OPEN cur1;
 
-            DELETE FROM suggestion;
             REPEAT
                 FETCH cur1 INTO user_help;
 
                 IF NOT done THEN
+                    DELETE FROM suggestion where user_id=user_help;
                     DELETE FROM suggestion_pod;
                     DELETE FROM suggestion_user;
 select user_help;
@@ -180,3 +180,57 @@ CREATE TABLE episode_log (
     UNIQUE (user_id, episode_id, timestamp)
 );
 
+ALTER TABLE toplist ADD COLUMN old_place INTEGER DEFAULT 0;
+
+DELIMITER $$
+DROP PROCEDURE IF EXISTS update_toplist $$
+CREATE PROCEDURE update_toplist()
+BEGIN
+    DECLARE deadlock INT DEFAULT 0;
+    DECLARE attempts INT DEFAULT 0;
+
+    DROP TABLE IF EXISTS toplist_temp;
+    CREATE TABLE toplist_temp (
+            podcast_id INT PRIMARY KEY REFERENCES podcast (id),
+            subscription_count INT NOT NULL DEFAULT 0,
+            old_place INT DEFAULT 0,
+            INDEX(podcast_id)
+    );
+
+    try_loop:WHILE (attempts<3) DO
+    BEGIN
+        DECLARE deadlock_detected CONDITION FOR 1213;
+            DECLARE EXIT HANDLER FOR deadlock_detected
+                BEGIN
+                    ROLLBACK;
+                    SET deadlock=1;
+                END;
+            SET deadlock=0;
+
+            START TRANSACTION;
+            DELETE FROM toplist_temp;
+            INSERT INTO toplist_temp (SELECT a.podcast_id, COUNT(*) AS count_subscription, 
+                                       (select (id - (select min(id) from toplist) + 1) from toplist where podcast_id = a.podcast_id)
+                        FROM (SELECT DISTINCT podcast_id, user_id
+                            FROM public_subscription) a
+                        GROUP BY podcast_id);
+            DELETE FROM toplist;
+            INSERT INTO toplist (podcast_id, subscription_count, id, old_place) (SELECT podcast_id, subscription_count, NULL, old_place FROM toplist_temp
+                        ORDER BY subscription_count DESC LIMIT 100);
+
+            COMMIT;
+        END;
+        IF deadlock=0 THEN
+                LEAVE try_loop;
+            ELSE
+                SET attempts=attempts+1;
+            END IF;
+            END WHILE try_loop;
+
+        IF deadlock=1 THEN
+            call FAIL('Toplist is not updated!');
+        END IF;
+        DROP TABLE IF EXISTS toplist_temp;
+
+END $$
+DELIMITER ;
